refactor: junction, junctionMover 분리
