+++ src/test/java/nextstep/ladder/domain/JunctionTest.java
@@ -35,45 +35,6 @@ class JunctionTest {
         assertThat(junction.hasDown()).isTrue();
     }
 
-    @Test
-    void 이동할_곳이_없으면_제자리다() {
-        ParticipantName name = new ParticipantName("1");
-        Junction junction = createJunction();
-
-        Junction result = junction.moveToResult(name);
-
-        assertThat(result).isEqualTo(junction);
-    }
-
-    @Test
-    void 방문한_곳은_갈_수_없다() {
-        ParticipantName name = new ParticipantName("1");
-        Junction junction = createJunction();
-        Junction rightJunction = createJunction();
-
-        junction.connectRight(rightJunction);
-        rightJunction.visit(name);
-
-        Junction result = junction.moveToResult(name);
-
-        assertThat(result).isEqualTo(junction);
-    }
-
-    @Test
-    void 이동할_곳이_있다면_이동한다() {
-        ParticipantName name = new ParticipantName("1");
-        Junction junction = createJunction();
-        Junction rightJunction = createJunction();
-        Junction rightDownJunction = createJunction();
-
-        rightJunction.connectDown(rightDownJunction);
-        junction.connectRight(rightJunction);
-
-        Junction result = junction.moveToResult(name);
-
-        assertThat(result).isEqualTo(rightDownJunction);
-    }
-
     public static Junction createJunction() {
         Leg leg = new LegBuilder().build();
         return new Junction(leg);
